+++ app/javascript/controllers/add_element_controller.js
@@ -1,0 +1,10 @@
+import { Controller } from "@hotwired/stimulus"
+
+// Connects to data-controller="add-element"
+export default class extends Controller {
+  static targets = ["template", "append"]
+
+  add() {
+    this.appendTarget.appendChild(this.templateTarget.content.cloneNode(true))
+  }
+}