+++ app/javascript/controllers/remove_element_controller.js
@@ -1,0 +1,8 @@
+import { Controller } from "@hotwired/stimulus"
+
+// Connects to data-controller="remove-element"
+export default class extends Controller {
+  removeSelf() {
+    this.element.remove()
+  }
+}