+++ app/models/ability.rb
@@ -5,6 +5,7 @@ class Ability
 
   def initialize(user)
     can :manage, User, id: user.id
+    can :manage, :expiration_notification
     can :manage, Coupon, user: user
     can :manage, LoyaltyCard, user: user
     can :manage, :search