+++ app/models/user.rb
@@ -10,4 +10,19 @@ class User < ApplicationRecord
   def remember_me
     super.nil? ? true : super
   end
+
+  def expiration_notification_channels=(channels)
+    raise ArgumentError, "expiration_notification_channels must be an array" unless channels.is_a?(Array)
+
+    super(channels.compact_blank.join(","))
+  end
+
+  def expiration_notification_channels
+    super&.split(",") || []
+  end
+
+  def has_notification_settings?
+    expiration_notification_days.present? &&
+      expiration_notification_channels.present?
+  end
 end