+++ config/environments/development.rb
@@ -74,4 +74,6 @@ Rails.application.configure do
 
   # Raise error when a before_action's only/except options reference missing actions
   config.action_controller.raise_on_missing_callback_actions = true
+
+  config.action_controller.default_url_options[:port] = 3000
 end