+++ config/routes.rb
@@ -6,7 +6,9 @@ Rails.application.routes.draw do
   scope "/:locale" do
     get "/pages/:id", to: "pages#show", as: :page
 
-    resource :user, only: %i[show edit update destroy]
+    resource :user, only: %i[show edit update destroy] do
+      resource :expiration_notification, only: %i[edit update], module: :users
+    end
     resources :coupons
     resources :loyalty_cards
     resource :search, only: :show