+++ db/migrate/20240711221707_add_expiration_notification_settings_to_users.rb
@@ -1,0 +1,8 @@
+class AddExpirationNotificationSettingsToUsers < ActiveRecord::Migration[7.1]
+  def change
+    add_column :users, :expiration_notification_days, :integer
+    add_column :users, :expiration_notification_title, :string
+    add_column :users, :expiration_notification_body, :string
+    add_column :users, :expiration_notification_channels, :string
+  end
+end