+++ db/migrate/20240713221900_add_locale_to_users.rb
@@ -1,0 +1,5 @@
+class AddLocaleToUsers < ActiveRecord::Migration[7.1]
+  def change
+    add_column :users, :locale, :string, null: false, default: "en"
+  end
+end